requirements.txt
discord.py>=2.4,<3.0
python-dotenv>=1.1,<2.0
